Minimal free-standing Rust binary
Note that this is just a minimal example of a freestanding Rust binary.
This binary expects various things, for example that a stack is
initialized when the _start function is called. So for any real use of
such a binary, more steps are required.

Corresponds with https://os.phil-opp.com/freestanding-rust-binary/

Next we will convert this into a minimal operating system kernel.